requirements.txt
fusepy
pyyaml
stringcase